added 100 LC and refactored codebase
